+++ src/main/java/blackjack/domain/cardgame/CardDeck.java
@@ -1,4 +1,8 @@
-package blackjack.domain;
+package blackjack.domain.cardgame;
+
+import blackjack.domain.card.Card;
+import blackjack.domain.card.CardNumber;
+import blackjack.domain.card.CardShape;
 
 import java.util.Collections;
 import java.util.Stack;
@@ -8,6 +12,7 @@ public class CardDeck {
 
     public CardDeck() {
         final Stack<Card> deck = new Stack<>();
+        // TODO: 2중 for문 제거
         for (CardNumber cardNumber : CardNumber.values()) {
             for (CardShape cardShape : CardShape.values()) {
                 deck.push(new Card(cardNumber, cardShape));
