+++ src/main/java/blackjack/domain/cardgame/CardGame.java
@@ -1,4 +1,7 @@
-package blackjack.domain;
+package blackjack.domain.cardgame;
+
+import blackjack.domain.player.Dealer;
+import blackjack.domain.player.Player;
 
 import java.util.List;
