+++ src/main/java/blackjack/domain/cardgame/WinningStatus.java
@@ -1,4 +1,4 @@
-package blackjack.domain;
+package blackjack.domain.cardgame;
 
 public enum WinningStatus {
     WIN("승"),
