+++ src/test/java/blackjack/domain/cardgame/CardGameResultTest.java
@@ -1,13 +1,16 @@
-package blackjack.domain;
+package blackjack.domain.cardgame;
 
+import blackjack.domain.cardgame.CardGameResult;
+import blackjack.domain.cardgame.WinningStatus;
+import blackjack.domain.player.Player;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.Test;
 
 import java.util.LinkedHashMap;
 import java.util.Map;
 
-import static blackjack.domain.WinningStatus.LOSE;
-import static blackjack.domain.WinningStatus.WIN;
+import static blackjack.domain.cardgame.WinningStatus.LOSE;
+import static blackjack.domain.cardgame.WinningStatus.WIN;
 import static blackjack.fixture.PlayerFixture.player;
 
 class CardGameResultTest {
