+++ src/test/java/blackjack/domain/player/NameTest.java
@@ -1,5 +1,6 @@
-package blackjack.domain;
+package blackjack.domain.player;
 
+import blackjack.domain.player.Name;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
